extract name formatting and put in new class in tools
